# martin/src/srv/server.rs
use std::future::Future;
use std::pin::Pin;
use std::time::Duration;

#[cfg(all(feature = "webui", not(docsrs)))]
use actix_web::guard::fn_guard;
use actix_web::http::header::{CACHE_CONTROL, HeaderValue};
use actix_web::middleware::{DefaultHeaders, NormalizePath, TrailingSlash};
use actix_web::web::Data;
use actix_web::{App, HttpResponse, HttpServer, Responder, middleware, route, web};
use futures::TryFutureExt as _;
#[cfg(feature = "lambda")]
use lambda_web::{is_running_on_lambda, run_actix_on_lambda};
use tracing_actix_web::TracingLogger;

#[cfg(all(feature = "webui", not(docsrs)))]
use crate::config::args::WebUiMode;
#[cfg(feature = "_catalog")]
use crate::config::file::ServerState;
use crate::config::file::srv::{DEFAULT_KEEP_ALIVE, DEFAULT_LISTEN_ADDRESSES, SrvConfig};
use crate::srv::ServerStartError;
#[cfg(any(not(feature = "webui"), docsrs))]
use crate::srv::admin::get_index_no_ui;
use crate::srv::admin::{Catalog, get_catalog};
#[cfg(all(feature = "webui", not(docsrs)))]
use crate::srv::admin::{get_index_ui_disabled, webui};
#[cfg(feature = "fonts")]
use crate::srv::fonts;
#[cfg(feature = "sprites")]
use crate::srv::sprites;
#[cfg(feature = "styles")]
use crate::srv::styles;
#[cfg(all(feature = "rendering", target_os = "linux"))]
use crate::srv::styles_rendering;
#[cfg(all(feature = "rendering", target_os = "linux"))]
use crate::srv::styles_static;
#[cfg(feature = "_tiles")]
use crate::srv::tiles;

/// List of keywords that cannot be used as source IDs. Some of these are reserved for future use.
///
/// Reserved keywords must never end in a "dot number" (e.g. ".1").
/// This list is documented in the `docs/content/using.md` file, which should be kept in sync.
pub const RESERVED_KEYWORDS: &[&str] = &[
    "_", "catalog", "config", "font", "health", "help", "index", "manifest", "metrics", "refresh",
    "reload", "sprite", "status",
];

/// Maps any classified error onto an HTTP response.
///
/// The status comes from the error's own `ErrorKind`, so a failure mode is classified once
/// where it is defined rather than at each handler that can surface it. Only failures the
/// caller cannot act on are logged.
#[cfg(any(feature = "_tiles", feature = "fonts", feature = "sprites"))]
pub fn map_error<E: std::fmt::Display + martin_core::Classify + ?Sized>(e: &E) -> actix_web::Error {
    use actix_web::error::{
        ErrorBadRequest, ErrorInternalServerError, ErrorNotFound, ErrorServiceUnavailable,
    };
    use martin_core::ErrorKind::{Internal, InvalidInput, NotFound, Unavailable};

    match e.kind() {
        NotFound => ErrorNotFound(e.to_string()),
        InvalidInput => ErrorBadRequest(e.to_string()),
        Unavailable => {
            tracing::error!("{e}");
            ErrorServiceUnavailable(e.to_string())
        }
        Internal => {
            tracing::error!("{e}");
            ErrorInternalServerError(e.to_string())
        }
    }
}

/// Helper struct for debounced warning messages in redirect handlers.
/// Ensures warnings are logged no more than once per hour to avoid log spam.
#[cfg(feature = "_catalog")]
pub struct DebouncedWarning {
    last_warning: std::sync::LazyLock<tokio::sync::Mutex<std::time::Instant>>,
}

#[cfg(feature = "_catalog")]
impl DebouncedWarning {
    /// Create a new `DebouncedWarning` instance
    pub const fn new() -> Self {
        Self {
            last_warning: std::sync::LazyLock::new(|| {
                tokio::sync::Mutex::new(std::time::Instant::now())
            }),
        }
    }

    /// Execute the provided closure at most once per hour.
    /// This allows tracing's log filtering to work correctly by keeping the warn! call site
    /// in the caller's context.
    pub async fn once_per_hour<F: FnOnce()>(&self, f: F) {
        let mut last = self.last_warning.lock().await;
        if last.elapsed() >= Duration::from_hours(1) {
            *last = std::time::Instant::now();
            drop(last);
            f();
        }
    }
}

/// Return 200 OK if healthy. Used for readiness and liveness probes.
#[cfg_attr(
    feature = "unstable-schemas",
    utoipa::path(
        get,
        path = "/health",
        responses((status = 200, description = "Service healthy", body = String)),
    )
)]
#[route("/health", method = "GET", method = "HEAD")]
pub async fn get_health() -> impl Responder {
    HttpResponse::Ok()
        .insert_header((CACHE_CONTROL, "no-cache"))
        .message_body("OK")
}

pub fn router(cfg: &mut web::ServiceConfig, usr_cfg: &SrvConfig) {
    // If route_prefix is configured, wrap all routes in a scope
    if let Some(prefix) = &usr_cfg.route_prefix {
        cfg.service(web::scope(prefix).configure(|cfg| {
            register_services(
                cfg,
                #[cfg(any(all(feature = "webui", not(docsrs)), feature = "_tiles"))]
                usr_cfg,
            );
        }));
        cfg.service(get_health);
    } else {
        register_services(
            cfg,
            #[cfg(any(all(feature = "webui", not(docsrs)), feature = "_tiles"))]
            usr_cfg,
        );
    }
}

/// Helper function to register all services
fn register_services(
    cfg: &mut web::ServiceConfig,
    #[cfg(any(all(feature = "webui", not(docsrs)), feature = "_tiles"))] usr_cfg: &SrvConfig,
) {
    cfg.service(get_health).service(get_catalog);

    #[cfg(feature = "_tiles")]
    {
        // Register tile format suffix redirects BEFORE the main tile route
        // because Actix-Web matches routes in registration order
        cfg.service(tiles::content::redirect_tile_ext)
            .service(tiles::metadata::get_source_info)
            .service(tiles::content::get_tile);

        // Register /tiles/ prefix redirect after main tile route
        cfg.service(tiles::content::redirect_tiles);

        if usr_cfg
            .endpoints
            .as_ref()
            .is_some_and(|endpoints| endpoints.purge_cache.unwrap_or(false))
        {
            cfg.service(super::cache::purge_source);
        }
    }

    #[cfg(feature = "sprites")]
    cfg.service(sprites::get_sprite_sdf_json)
        .service(sprites::redirect_sdf_sprites_json)
        .service(sprites::get_sprite_json)
        .service(sprites::redirect_sprites_json)
        .service(sprites::get_sprite_sdf_png)
        .service(sprites::redirect_sdf_sprites_png)
        .service(sprites::get_sprite_png)
        .service(sprites::redirect_sprites_png);

    #[cfg(feature = "fonts")]
    // Register the glyph file-extension redirect BEFORE the main font route
    cfg.service(fonts::redirect_font_ext)
        .service(fonts::get_font)
        .service(fonts::redirect_fonts);

    #[cfg(feature = "styles")]
    cfg.service(styles::get_style_json)
        .service(styles::redirect_styles);

    // `.jpeg` redirects must register before the main routes so they win.
    #[cfg(all(feature = "rendering", target_os = "linux"))]
    cfg.service(styles_static::redirect_static_jpeg)
        .service(styles_static::get_rendered_static_style)
        .service(styles_static::post_rendered_static_style)
        .service(styles_rendering::redirect_tile_jpeg)
        .service(styles_rendering::get_rendered_tile_style);

    #[cfg(all(feature = "webui", not(docsrs)))]
    match usr_cfg.web_ui.unwrap_or_default() {
        WebUiMode::EnableForAll => {
            cfg.service(actix_web_static_files::ResourceFiles::new(
                "/",
                webui::generate(),
            ));
        }
        WebUiMode::Enable => {
            let guard = fn_guard(|c| c.head()
        .peer_addr
        .is_some_and(|addr| addr.ip().is_loopback()));
            cfg.service(
                web::scope("")
                    .guard(guard)
                    .service(actix_web_static_files::ResourceFiles::new(
                        "/",
                        webui::generate(),
                    )),
            )
            .service(get_index_ui_disabled);
        }
        WebUiMode::Disable => {
            cfg.service(get_index_ui_disabled);
        }
    }

    #[cfg(any(not(feature = "webui"), docsrs))]
    cfg.service(get_index_no_ui);
}

type Server = Pin<Box<dyn Future<Output = Result<(), ServerStartError>>>>;

fn cache_control_middleware(value: Option<HeaderValue>) -> middleware::Condition<DefaultHeaders> {
    let enabled = value.is_some();
    let value = value.unwrap_or_else(|| HeaderValue::from_static(""));
    middleware::Condition::new(enabled, DefaultHeaders::new().add((CACHE_CONTROL, value)))
}

/// Create a future for an Actix web server together with the listening address.
#[hotpath::measure]
pub fn new_server(
    config: SrvConfig,
    #[cfg(feature = "_catalog")] state: ServerState,
) -> Result<(Server, String), ServerStartError> {
    let cache_control = config.cache_control_header();
    #[cfg(feature = "metrics")]
    let prometheus = {
        let metrics_endpoint = if let Some(prefix) = &config.route_prefix {
            format!("{prefix}/_/metrics")
        } else {
            "/_/metrics".to_owned()
        };
        actix_web_prom::PrometheusMetricsBuilder::new("martin")
            .registry(prometheus::default_registry().clone())
            .endpoint(&metrics_endpoint)
            // `endpoint="UNKNOWN"` instead of `endpoint="/foo/bar"`
            .mask_unmatched_patterns("UNKNOWN")
            .const_labels(
                config
                    .observability
                    .clone()
                    .unwrap_or_default()
                    .metrics
                    .unwrap_or_default()
                    .add_labels,
            )
            .build()
            .map_err(ServerStartError::MetricsInitialisation)?
    };
    let catalog = Catalog::new(
        #[cfg(any(feature = "sprites", feature = "fonts", feature = "styles"))]
        &state,
    )?;

    let keep_alive = Duration::from_secs(config.keep_alive.unwrap_or(DEFAULT_KEEP_ALIVE));
    let worker_processes = config.worker_processes.unwrap_or_else(num_cpus::get);
    let listen_addresses = config
        .listen_addresses
        .clone()
        .unwrap_or_else(|| DEFAULT_LISTEN_ADDRESSES.to_owned());

    let cors_config = config.cors.clone().unwrap_or_default();
    cors_config.validate()?;
    cors_config.log_current_configuration();

    let factory = move || {
        let cors_middleware = cors_config.make_cors_middleware();

        let app = App::new()
            .app_data(Data::new(catalog.clone()))
            .app_data(Data::new(config.clone()));

        #[cfg(feature = "_tiles")]
        let app = app.app_data(Data::new(state.tile_manager.clone()));

        #[cfg(feature = "sprites")]
        let app = app
            .app_data(Data::new(state.sprites.clone()))
            .app_data(Data::new(state.sprite_cache.clone()));

        #[cfg(feature = "fonts")]
        let app = app
            .app_data(Data::new(state.fonts.clone()))
            .app_data(Data::new(state.font_cache.clone()));

        #[cfg(feature = "styles")]
        let app = app.app_data(Data::new(state.styles.clone()));

        let app = app.wrap(middleware::Condition::new(
            cors_middleware.is_some(),
            cors_middleware.unwrap_or_default(),
        ));

        #[cfg(feature = "metrics")]
        let app = app.wrap(prometheus.clone());

        #[cfg(feature = "tui")]
        let app = app.wrap(middleware::Condition::new(
            crate::tui::is_installed(),
            middleware::from_fn(crate::tui::observe),
        ));

        app.wrap(TracingLogger::default())
            .wrap(cache_control_middleware(cache_control.clone()))
            .wrap(NormalizePath::new(TrailingSlash::MergeOnly))
            .configure(|c| router(c, &config))
    };

    #[cfg(feature = "lambda")]
    if is_running_on_lambda() {
        let server = run_actix_on_lambda(factory).map_err(ServerStartError::Lambda);
        return Ok((Box::pin(server), "(aws lambda)".into()));
    }

    let server = HttpServer::new(factory)
        .bind(listen_addresses.clone())
        .map_err(|e| ServerStartError::Binding(e, listen_addresses.clone()))?;

    let listen_addresses = server
        .addrs()
        .first()
        .map_or(listen_addresses, ToString::to_string);

    let server = server
        .keep_alive(keep_alive)
        .shutdown_timeout(0)
        .workers(worker_processes)
        .run()
        .err_into();

    Ok((Box::pin(server), listen_addresses))
}